// src/screen/espaj/spaj-physical-habits/spaj-physical-habits.type.ts
export type defaultTFormPhysicalHabits = {
  height: '';
  weight: '';
  smokerStatus: '';
  averageCigarettes: '';
  durationtotalAlcohol: '';
  averageTotalAlcohol: '';
  totalAlcoholconsumed: '';
  questionOfMedicine: '';
  medicineName: '';
};

export type TFormPhysicalHabits = {
  height: string;
  weight: string;
  smokerStatus: string;
  averageCigarettes: string;
  durationtotalAlcohol: string;
  averageTotalAlcohol: string;
  totalAlcoholconsumed: string;
  questionOfMedicine: string;
  medicineName: string;
};